GV depreciated is now gone. Tests are Old and need an update.
