+++ ndn_distributed_repo/global_view/__init__.py
@@ -1,3 +1,1 @@
-from .global_view import * 
-from .file_metadata import * 
-from .node_metadata import *
+from .global_view import GlobalView
