--- ndn_distributed_repo/global_view_2/__init__.py
@@ -1,1 +1,0 @@
-from .global_view import GlobalView
