+++ ndn_distributed_repo/handle_data_storage/data_storage_handle.py
@@ -1,5 +1,3 @@
-
-
 import asyncio as aio
 from hashlib import blake2b
 from typing import Dict
